@@ -5,13 +5,13 @@ publish_to: "none"
 version: 1.0.0+1
 
 environment:
-  sdk: ">=3.3.1 <4.0.0"
+  sdk: ">=3.4.0 <4.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  cupertino_icons: ^1.0.6
-  flutter_native_splash: ^2.0.4
+  cupertino_icons: ^1.0.8
+  flutter_native_splash: ^2.4.0
   flutter_svg: ^2.0.10+1
   qr_flutter: ^4.1.0
 
